feat: implement Scanner with Iter and Iter2 methods; add corresponding tests in helper.go
